refactor(core): remove ReportGenerator in favor of SessionStore pipeline (#2142)
ReportGenerator was no longer used by the Agent class — it had been
fully replaced by SessionStore + exportSessionReport. This removes the
dead code and updates the two test files that used ReportGenerator as a
helper to produce test reports for ReportMergingTool tests.

- Delete report-generator.ts and its dedicated test file
- Remove IReportGenerator, ReportGenerator, nullReportGenerator exports
- Update report-merge-count.test.ts to use reportHTMLContent directly
- Update merge-browser-parse.test.ts to write directory-mode reports
  manually (screenshots + HTML) instead of via ReportGenerator
