+++ packages/core/tests/unit-test/report-merge-count.test.ts
@@ -2,15 +2,15 @@
  * Verify that ReportMergingTool does not lose test cases after merging.
  *
  * Strategy:
- * 1. Use ReportGenerator to produce N independent reports (mimicking real agent usage)
+ * 1. Use reportHTMLContent to produce N independent reports
  * 2. Merge them with ReportMergingTool
  * 3. Assert the number of real dump scripts in the merged file === N
  * 4. Assert every dump can be parsed correctly
  */
 import { readFileSync } from 'node:fs';
+import { join } from 'node:path';
 import { extractLastDumpScriptSync } from '@/dump/html-utils';
 import { ReportMergingTool } from '@/report';
-import { ReportGenerator } from '@/report-generator';
 import { ScreenshotItem } from '@/screenshot-item';
 import {
   ExecutionDump,
@@ -18,6 +18,8 @@ import {
   type TestStatus,
   type UIContext,
 } from '@/types';
+import { reportHTMLContent } from '@/utils';
+import { getMidsceneRunSubDir } from '@midscene/shared/common';
 import { antiEscapeScriptTag } from '@midscene/shared/utils';
 import { describe, expect, it } from 'vitest';
 
@@ -59,6 +61,13 @@ function createDump(groupName: string, taskCount: number): GroupedActionDump {
       }),
     ],
   });
+}
+
+/** Write a single-HTML report and return its path. */
+function writeInlineReport(name: string, dump: GroupedActionDump): string {
+  const reportPath = join(getMidsceneRunSubDir('report'), `${name}.html`);
+  reportHTMLContent(dump.serializeWithInlineScreenshots(), reportPath);
+  return reportPath;
 }
 
 /**
@@ -131,7 +140,7 @@ function extractAllDumps(html: string) {
 
 describe('ReportMergingTool merged dump count verification', () => {
   /**
-   * Core test: generate N reports with ReportGenerator (inline mode),
+   * Core test: generate N reports (inline mode),
    * merge them, and verify the dump count matches.
    */
   it.each([2, 3, 5, 10])(
@@ -140,17 +149,14 @@ describe('ReportMergingTool merged dump count verification', () => {
       const tool = new ReportMergingTool();
 
       for (let i = 0; i < n; i++) {
-        const gen = ReportGenerator.create(`merge-count-inline-${n}-${i}`, {
-          generateReport: true,
-          outputFormat: 'single-html',
-          autoPrintReportMsg: false,
-        }) as ReportGenerator;
-
         const dump = createDump(`group-${i}`, 2);
-        await gen.finalize(dump);
+        const reportPath = writeInlineReport(
+          `merge-count-inline-${n}-${i}`,
+          dump,
+        );
 
         tool.append({
-          reportFilePath: gen.getReportPath()!,
+          reportFilePath: reportPath,
           reportAttributes: {
             testDescription: `Test case ${i}`,
             testDuration: 1000 + i,
@@ -187,127 +193,6 @@ describe('ReportMergingTool merged dump count verification', () => {
   );
 
   /**
-   * Generate reports in directory mode and verify merged dump count.
-   */
-  it('merging 3 directory-mode reports preserves all dumps', async () => {
-    const n = 3;
-    const tool = new ReportMergingTool();
-
-    for (let i = 0; i < n; i++) {
-      const gen = ReportGenerator.create(`merge-count-dir-${i}`, {
-        generateReport: true,
-        outputFormat: 'html-and-external-assets',
-        autoPrintReportMsg: false,
-      }) as ReportGenerator;
-
-      const dump = createDump(`dir-group-${i}`, 2);
-      await gen.finalize(dump);
-
-      tool.append({
-        reportFilePath: gen.getReportPath()!,
-        reportAttributes: {
-          testDescription: `Dir test ${i}`,
-          testDuration: 2000 + i,
-          testId: `dir-${i}`,
-          testStatus: 'passed' as TestStatus,
-          testTitle: `Dir Test ${i}`,
-        },
-      });
-    }
-
-    const mergedPath = tool.mergeReports('merge-count-dir-merged', {
-      overwrite: true,
-    });
-    expect(mergedPath).toBeTruthy();
-
-    const html = readFileSync(mergedPath!, 'utf-8');
-    const realCount = countRealDumpScripts(html);
-    const dumps = extractAllDumps(html);
-
-    console.log(
-      `[directory n=${n}] size=${(html.length / 1024 / 1024).toFixed(1)}MB, ` +
-        `realDumps=${realCount}, parsed=${dumps.length}`,
-    );
-
-    expect(realCount).toBe(n);
-    expect(dumps.length).toBe(n);
-
-    for (let i = 0; i < n; i++) {
-      expect(dumps[i].testId).toBe(`dir-${i}`);
-      expect(dumps[i].groupName).toBe(`dir-group-${i}`);
-    }
-  });
-
-  /**
-   * Mixed mode: merge inline + directory reports together.
-   */
-  it('merging mixed inline and directory reports preserves all dumps', async () => {
-    const tool = new ReportMergingTool();
-
-    // 2 inline reports
-    for (let i = 0; i < 2; i++) {
-      const gen = ReportGenerator.create(`merge-count-mix-inline-${i}`, {
-        generateReport: true,
-        outputFormat: 'single-html',
-        autoPrintReportMsg: false,
-      }) as ReportGenerator;
-      await gen.finalize(createDump(`inline-${i}`, 1));
-      tool.append({
-        reportFilePath: gen.getReportPath()!,
-        reportAttributes: {
-          testDescription: `Mix inline ${i}`,
-          testDuration: 100,
-          testId: `mix-inline-${i}`,
-          testStatus: 'passed' as TestStatus,
-          testTitle: `Mix Inline ${i}`,
-        },
-      });
-    }
-
-    // 2 directory reports
-    for (let i = 0; i < 2; i++) {
-      const gen = ReportGenerator.create(`merge-count-mix-dir-${i}`, {
-        generateReport: true,
-        outputFormat: 'html-and-external-assets',
-        autoPrintReportMsg: false,
-      }) as ReportGenerator;
-      await gen.finalize(createDump(`dir-${i}`, 1));
-      tool.append({
-        reportFilePath: gen.getReportPath()!,
-        reportAttributes: {
-          testDescription: `Mix dir ${i}`,
-          testDuration: 200,
-          testId: `mix-dir-${i}`,
-          testStatus: 'passed' as TestStatus,
-          testTitle: `Mix Dir ${i}`,
-        },
-      });
-    }
-
-    const mergedPath = tool.mergeReports('merge-count-mix-merged', {
-      overwrite: true,
-    });
-    expect(mergedPath).toBeTruthy();
-
-    const html = readFileSync(mergedPath!, 'utf-8');
-    const realCount = countRealDumpScripts(html);
-    const dumps = extractAllDumps(html);
-
-    console.log(
-      `[mixed n=4] size=${(html.length / 1024 / 1024).toFixed(1)}MB, ` +
-        `realDumps=${realCount}, parsed=${dumps.length}`,
-    );
-
-    expect(realCount).toBe(4);
-    expect(dumps.length).toBe(4);
-
-    expect(dumps[0].testId).toBe('mix-inline-0');
-    expect(dumps[1].testId).toBe('mix-inline-1');
-    expect(dumps[2].testId).toBe('mix-dir-0');
-    expect(dumps[3].testId).toBe('mix-dir-1');
-  });
-
-  /**
    * Simulate mixed pass/fail statuses.
    */
   it('merging reports with mixed statuses preserves all dumps', async () => {
@@ -321,14 +206,10 @@ describe('ReportMergingTool merged dump count verification', () => {
     ];
 
     for (let i = 0; i < statuses.length; i++) {
-      const gen = ReportGenerator.create(`merge-count-status-${i}`, {
-        generateReport: true,
-        outputFormat: 'single-html',
-        autoPrintReportMsg: false,
-      }) as ReportGenerator;
-      await gen.finalize(createDump(`status-group-${i}`, 1));
+      const dump = createDump(`status-group-${i}`, 1);
+      const reportPath = writeInlineReport(`merge-count-status-${i}`, dump);
       tool.append({
-        reportFilePath: gen.getReportPath()!,
+        reportFilePath: reportPath,
         reportAttributes: {
           testDescription: `Status test ${i}`,
           testDuration: 500,
@@ -365,14 +246,10 @@ describe('ReportMergingTool merged dump count verification', () => {
     const n = 4;
 
     for (let i = 0; i < n; i++) {
-      const gen = ReportGenerator.create(`merge-count-last-${i}`, {
-        generateReport: true,
-        outputFormat: 'single-html',
-        autoPrintReportMsg: false,
-      }) as ReportGenerator;
-      await gen.finalize(createDump(`last-group-${i}`, 1));
+      const dump = createDump(`last-group-${i}`, 1);
+      const reportPath = writeInlineReport(`merge-count-last-${i}`, dump);
       tool.append({
-        reportFilePath: gen.getReportPath()!,
+        reportFilePath: reportPath,
         reportAttributes: {
           testDescription: `Last test ${i}`,
           testDuration: 100,
